creating state mgmt layer module
