+++ apps/backend/pits_n_giggles.py
@@ -35,7 +35,7 @@ from colorama import Fore, Style, init
 
 from apps.backend.common.config import load_config
 from apps.backend.common.png_logger import initLogger
-from apps.backend.state_mgmt_layer.telemetry_state import initSessionState
+from apps.backend.state_mgmt_layer import initStateManagementLayer
 from apps.backend.telemetry_layer import setupForwarder, setupTelemetryTask
 from apps.backend.ui_intf_layer import (TelemetryWebServer,
                                         initTelemetryWebServer)
@@ -180,10 +180,7 @@ async def main(args: argparse.Namespace) -> None:
     png_logger.info(f"Python version {sys.version}")
     png_logger.info(config)
 
-    initSessionState(
-        logger=png_logger,
-        process_car_setups=config.process_car_setup
-    )
+    initStateManagementLayer(logger=png_logger, process_car_setups=config.process_car_setup)
 
     tasks: List[asyncio.Task] = []
