+++ apps/backend/state_mgmt_layer/state_layer_init.py
@@ -1,0 +1,35 @@
+# MIT License
+#
+# Copyright (c) [2024] [Ashwin Natarajan]
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+# ------------------------- IMPORTS ------------------------------------------------------------------------------------
+
+import logging
+
+from .telemetry_state import initSessionState
+from .telemetry_web_api import initApiLayer
+
+# -------------------------------------- FUNCTIONS ---------------------------------------------------------------------
+
+def initStateManagementLayer(logger: logging.Logger, process_car_setups: bool) -> None:
+    """Initialise the state management layer"""
+    initSessionState(logger=logger, process_car_setups=process_car_setups)
+    initApiLayer(logger=logger)
